Added flow step for existing wallet onboarding
