+++ packages/commonwealth/client/scripts/views/modals/AuthModal/common/ModalBase/EVMWalletsSubModal/EVMWalletsSubModal.scss
@@ -1,4 +1,4 @@
-@import '../../../../../styles/shared';
+@import '../../../../../../../styles/shared';
 
 .EVMWalletsSubModal {
   background-color: black;
