+++ packages/commonwealth/client/scripts/views/modals/AuthModal/common/ModalBase/EmailForm/EmailForm.scss
@@ -1,4 +1,4 @@
-@import '../../../../../styles/shared';
+@import '../../../../../../../styles/shared';
 
 .EmailForm {
   display: flex;
