+++ packages/commonwealth/client/scripts/views/modals/AuthModal/common/ModalBase/EmailForm/EmailForm.tsx
@@ -1,8 +1,8 @@
 import CWCircleMultiplySpinner from 'client/scripts/views/components/component_kit/new_designs/CWCircleMultiplySpinner';
 import React from 'react';
-import { CWButton } from '../../../components/component_kit/new_designs/CWButton';
-import { CWForm } from '../../../components/component_kit/new_designs/CWForm';
-import { CWTextInput } from '../../../components/component_kit/new_designs/CWTextInput';
+import { CWButton } from '../../../../../components/component_kit/new_designs/CWButton';
+import { CWForm } from '../../../../../components/component_kit/new_designs/CWForm';
+import { CWTextInput } from '../../../../../components/component_kit/new_designs/CWTextInput';
 import './EmailForm.scss';
 import { emailValidationSchema } from './validation';
