+++ packages/commonwealth/client/scripts/views/modals/AuthModal/common/ModalBase/EmailForm/validation.ts
