+++ packages/commonwealth/client/scripts/views/modals/AuthModal/common/ModalBase/MobileWalletConfirmationSubModal/MobileWalletConfirmationSubModal.scss
@@ -1,6 +1,6 @@
-@import '../../../../../../styles/shared';
+@import '../../../../../../../styles/shared';
 
-.ModalLayout {
+.MobileWalletConfirmationSubModal {
   display: flex;
   flex-direction: column;
   position: relative;
@@ -31,17 +31,13 @@
 
   .header {
     text-align: center;
-    padding: 0px 12px;
-  }
-
-  .description {
-    text-align: center;
-    color: $neutral-600;
-    margin-top: -12px;
-    padding: 0px 12px;
   }
 
   .content {
+    .header {
+      color: $neutral-600;
+    }
+
     padding: 0 48px;
 
     @include extraSmall {
@@ -56,29 +52,11 @@
       padding-bottom: 32px;
     }
 
-    display: flex;
-    flex-direction: column;
     justify-content: center;
 
     .Text {
       display: initial;
-    }
-
-    a,
-    a:visited,
-    a:active,
-    a:hover,
-    button {
-      color: $primary-600;
-    }
-
-    button {
-      text-decoration: none;
-      outline: none;
-      padding: 0;
-      margin: 0;
-      border: 0;
-      cursor: pointer;
+      color: $neutral-400;
     }
   }
 }
