+++ packages/commonwealth/client/scripts/views/modals/AuthModal/common/ModalBase/MobileWalletConfirmationSubModal/MobileWalletConfirmationSubModal.tsx
@@ -1,12 +1,12 @@
 import { CWButton } from 'client/scripts/views/components/component_kit/new_designs/CWButton';
 import React from 'react';
-import { CWIcon } from '../../../components/component_kit/cw_icons/cw_icon';
-import { CWText } from '../../../components/component_kit/cw_text';
+import { CWIcon } from '../../../../../components/component_kit/cw_icons/cw_icon';
+import { CWText } from '../../../../../components/component_kit/cw_text';
 import {
   CWModal,
   CWModalBody,
   CWModalFooter,
-} from '../../../components/component_kit/new_designs/CWModal';
+} from '../../../../../components/component_kit/new_designs/CWModal';
 import './MobileWalletConfirmationSubModal.scss';
 
 type MobileWalletConfirmationSubModal = {
