+++ packages/commonwealth/client/scripts/views/modals/AuthModal/common/ModalBase/MobileWalletConfirmationSubModal/index.ts
