+++ packages/commonwealth/client/scripts/views/modals/AuthModal/common/ModalBase/index.ts
@@ -1,0 +1,1 @@
+export { ModalBase } from './ModalBase';
