--- packages/commonwealth/client/scripts/views/modals/AuthModal/common/ModalLayout/index.ts
@@ -1,1 +1,0 @@
-export { ModalLayout } from './ModalLayout';
